Graph implemented with buttons on screen(however it is very laggy)
I have added the graph but the dots dont show up properly yet and I'll fix that in the coming week. I had to comment out the newly added heatmap stuff while working on it as the code would not run otherwise. The heatmap stuff is no longer commented out.